Ensure that pullToRefreshView stop observing scrollView properties.
Since the SVPullToRefreshView holds a weak reference to the scrollView,
it is possible for the scrollView to be nil by the time dealloc is called.
This situation leaks Key-Value Observers, triggering a runtime warning,
and leading to a potential crash.
    
These changes attempt to ensure that scrollView observers are added & removed
whenever the pullToRefresh view changes, and also during init/dealloc.