Preserve existing tableFooterView when configuring infinite scrolling support.
The current infinite scrolling implementation includes a toggle for hiding
the SVPullToRefresh view by setting the tableFooterView to nil. However, it
is possible for a tableView to have an existing footerView that will be
removed by this procedure.  A common use case is setting the tableFooterView
to an empty view (frame = CGRectZero) in order to prevent extra empty cells
from appearing in the tableView.

This changeset saves the original tableFooterView in an ivar, and restores
it when the 'showsInfiniteScrolling' property is set to NO.